% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_init.R
\name{devoirs_init}
\alias{devoirs_init}
\title{Initialize a Quarto/HTML assignment document}
\usage{
devoirs_init(clear_cache = TRUE)
}
\arguments{
\item{clear_cache}{If \code{TRUE} will add a button to reset the document's
devoirs answer cache}
}
\description{
Starts the ball rolling on a new assignment document. Creates the button
that, when pressed by the reader of the assignment, will collect of the the
answers to \code{{devoirs}} items.
}
