% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_knitr_hooks.R
\name{devoirs_knitr_hooks}
\alias{devoirs_knitr_hooks}
\title{For setting the knitr hooks}
\usage{
devoirs_knitr_hooks(x)
}
\description{
This should \emph{not} be called directly by the user.
}
\details{
See zzz.R for where knitr_hooks$set() is actually invoked
}
