% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_mc_dropdown.R
\name{devoirs_mc_dropdown}
\alias{devoirs_mc_dropdown}
\title{An inline multiple choice question in dropdown menu format}
\usage{
devoirs_mc_dropdown(
  label,
  choices,
  correct,
  points = 1,
  devoirs_feedback = knitr::opts_chunk$get("devoirs_feedback")
)
}
\description{
Use this in an inline chunk.
}
