% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_mc_engine.R
\name{devoirs_mc_engine}
\alias{devoirs_mc_engine}
\title{Defines a knitr engine for multiple choice questions.}
\usage{
devoirs_mc_engine(options)
}
\description{
This should \emph{not} be called directly by the user.
}
