% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_num.R
\name{devoirs_num}
\alias{devoirs_num}
\title{Add a widget that asks for numerical input}
\usage{
devoirs_num(
  label,
  true_answer,
  interval = 0,
  devoirs_feedback = knitr::opts_chunk$get("devoirs_feedback"),
  placeholder = "num...",
  width = 6,
  points = 1
)
}
\description{
Use this in an inline chunk
}
