% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_text_engine.R
\name{devoirs_text_engine}
\alias{devoirs_text_engine}
\title{Knitr engine devoirs_text for adding a free text entry widget to the document.}
\usage{
devoirs_text_engine(options)
}
\description{
This should \emph{not} be called directly by the user.
}
