% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_pdf.R
\name{devoirs_word_pdf}
\alias{devoirs_word_pdf}
\title{Versions of the devoirs inline widgets suited for PDF output}
\usage{
devoirs_word_pdf(
  id,
  placeholder = "",
  width = 10,
  true_answer = "",
  devoirs_feedback = TRUE,
  points = 1
)
}
\description{
NOTE IN DRAFT: Go back and rename the original \code{devoirs_set()} and
similar functions to \code{devoirs_set_html()} and such.
Then, write your own \code{devoirs_set()} that will look at
the target document format and
choose either the \verb{_html} or \verb{_pdf} version as appropriate.
}
