% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{process_math_markdown}
\alias{process_math_markdown}
\title{Internals for the devoirs2 package}
\usage{
process_math_markdown(text_string)
}
\description{
These functions are invoked automatically during the construction of
a Quarto/devoirs2 document. They provide the necessary JavaScript functionality
}
\details{
Replaces the LaTeX dollar signs in a string to be displayed as feedback
so that MathJax will process the string.
}
