% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_word.R
\name{devoirs_word}
\alias{devoirs_word}
\title{Add a widget suited for a single word}
\usage{
devoirs_word(
  id,
  placeholder = "",
  width = 10,
  true_answer = "",
  devoirs_feedback = TRUE,
  points = 1
)
}
\description{
Use this as an inline function
}
